- feat(frameworks) supports expressjs, koajs, nestjs framework deployment - feat(cli-service) optimize the structure of the output directory structure

+++ plugins/node-plugin/src/hooks/build.ts
@@ -1,7 +1,7 @@
 import { ConfigUtil } from '@malagu/cli-common';
 import { BuildContext, ConfigurationContext } from '@malagu/cli-service';
 import { join } from 'path';
-import { renameSync, writeFileSync } from 'fs-extra';
+import { renameSync, writeFileSync, existsSync } from 'fs-extra';
 
 const entryContent =  `
 const app = require('./_index');
@@ -9,16 +9,17 @@ let target = app;
 if (typeof app === 'object' && app.default) {
     target = app.default;
 }
-if (typeof target === 'function') {
-    if (typeof target.listen === 'function') {
-        const server = target.listen(PORT);
-        if (typeof server === 'object') {
-            server.timeout = 0;
-            server.keepAliveTimeout = 0;
-        }
-    } else {
-        target(PORT);
+if (target && typeof target.listen === 'function') {
+    const server = target.listen(PORT);
+    if (typeof server === 'object') {
+        server.timeout = 0;
+        server.keepAliveTimeout = 0;
     }
+    return;
+}
+
+if (typeof target === 'function') {
+    target(PORT);
 }
 `;
 
@@ -29,6 +30,8 @@ export default async (context: BuildContext) => {
     const port = ConfigUtil.getBackendMalaguConfig(cfg).server?.port || 9000;
     const oldIndexPath = join(outputPath, 'index.js');
     const newIndexPath = join(outputPath, '_index.js');
-    renameSync(oldIndexPath, newIndexPath);
-    writeFileSync(oldIndexPath, entryContent.replace(/PORT/g, port), { encoding: 'utf8' });
+    if (existsSync(oldIndexPath)) {
+        renameSync(oldIndexPath, newIndexPath);
+        writeFileSync(oldIndexPath, entryContent.replace(/PORT/g, port), { encoding: 'utf8' });
+    }
 };
